Stop truncating the _findfirst handle, which crashed x64 level loads
Loading any level without an existing local swap file killed the x64
build instantly: c0000005, faulting inside ntdll, with DebugLog ending
at "Image Loader:  <level>.grf".

bProcessLocalSwapFile enumerates the existing *.swp files to decide
which to evict, and stored the search handle in an int32:

    int32 i4_ffhandle = _findfirst(str_wildcard, &fnd);

_findfirst returns intptr_t. On Win32 that is 32 bits and the code is
correct by accident; on x64 it is 64 bits, so the assignment threw away
the top half of the handle and the CRT then dereferenced the remains in
_findnext/_findclose. Hence the fault inside ntdll rather than in our
own code, and hence page heap not moving it - nothing was overrunning a
buffer, we were handing out a wild pointer.

Widen it to intptr_t. Sole occurrence of this pattern in the tree.

x64 now creates the swap and loads the level: "Cleaning out old swap
files." / "Creating local swap file (Compressed)..." / "done creating
local swap file..." and Total unique meshes: 1101 for as2 - identical to
Win32 on the same level. Win32 Release still builds clean and is
unaffected (the type is 32-bit there either way).

+++ jp2_pc/Source/Lib/Loader/ImageLoader.cpp
@@ -578,7 +578,7 @@ bool CLoadImageDirectory::bProcessLocalSwapFile
 	strcpy(str_wildcard,str_local_swap_dir);
 	strcat(str_wildcard,"\\*.swp");
 
-	int32 i4_ffhandle = _findfirst(str_wildcard, &fnd);
+	intptr_t i4_ffhandle = _findfirst(str_wildcard, &fnd);
 
 	//
 	// If FindFirst reports an error there is probably no files that match so copy the
